fix: moved SidePanelController inside SidePanel component folder
