+++ frontend/pages/community/crm/companies.tsx
@@ -4,7 +4,7 @@ import ContentLayout from "~community/common/components/templates/ContentLayout/
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { IconName } from "~community/common/types/IconTypes";
 import CompanyModalController from "~community/crm/components/organisms/CompanyModalController/CompanyModalController";
-import CompanySidePanelController from "~community/crm/components/organisms/CompanySidePanelController/CompanySidePanelController";
+import CompanySidePanelController from "~community/crm/components/organisms/CompanySidePanel/CompanySidePanelController";
 import { CompanyTable } from "~community/crm/components/organisms/CompanyTable/CompanyTable";
 import { useCrmStore } from "~community/crm/store/store";
 import { CrmModalTypes } from "~community/crm/types/ModalTypes";
