+++ frontend/src/community/crm/components/organisms/CompanySidePanel/CompanySidePanelController.tsx
